+++ cmd/zetae2e/local/local.go
@@ -46,7 +46,7 @@ const (
 	flagTestV2Migration   = "test-v2-migration"
 	flagSkipTrackerCheck  = "skip-tracker-check"
 	flagSkipPrecompiles   = "skip-precompiles"
-	flagUpgradeGateways   = "upgrade-gateways"
+	flagUpgradeContracts  = "upgrade-contracts"
 )
 
 var (
@@ -84,7 +84,8 @@ func NewLocalCmd() *cobra.Command {
 	cmd.Flags().Bool(flagTestV2Migration, false, "set to true to run tests for v2 contracts migration test")
 	cmd.Flags().Bool(flagSkipTrackerCheck, false, "set to true to skip tracker check at the end of the tests")
 	cmd.Flags().Bool(flagSkipPrecompiles, false, "set to true to skip stateful precompiled contracts test")
-	cmd.Flags().Bool(flagUpgradeGateways, false, "set to true to upgrade gateways during setup for ZEVM and EVM")
+	cmd.Flags().
+		Bool(flagUpgradeContracts, false, "set to true to upgrade Gateways and ERC20Custody contracts during setup for ZEVM and EVM")
 
 	return cmd
 }
@@ -114,7 +115,7 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 		testV2            = must(cmd.Flags().GetBool(flagTestV2))
 		testV2Migration   = must(cmd.Flags().GetBool(flagTestV2Migration))
 		skipPrecompiles   = must(cmd.Flags().GetBool(flagSkipPrecompiles))
-		upgradeGateways   = must(cmd.Flags().GetBool(flagUpgradeGateways))
+		upgradeContracts  = must(cmd.Flags().GetBool(flagUpgradeContracts))
 	)
 
 	logger := runner.NewLogger(verbose, color.FgWhite, "setup")
@@ -329,7 +330,8 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 				e2etests.TestPrecompilesPrototypeName,
 				e2etests.TestPrecompilesPrototypeThroughContractName,
 				e2etests.TestPrecompilesStakingName,
-				e2etests.TestPrecompilesStakingThroughContractName,
+				// Disabled until further notice, check https://github.com/zeta-chain/node/issues/3005.
+				// e2etests.TestPrecompilesStakingThroughContractName,
 				e2etests.TestPrecompilesBankName,
 				e2etests.TestPrecompilesBankFailName,
 				e2etests.TestPrecompilesBankThroughContractName,
@@ -414,9 +416,8 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 		eg.Go(tonTestRoutine(conf, deployerRunner, verbose, tonTests...))
 	}
 
-	// upgrade gateways
-	if upgradeGateways {
-		deployerRunner.UpgradeGateways()
+	if upgradeContracts {
+		deployerRunner.UpgradeGatewaysAndERC20Custody()
 	}
 
 	if testV2 || testV2Migration {