+++ precompiles/types/errors_test.go
@@ -13,9 +13,7 @@ func Test_ErrInvalidAddr(t *testing.T) {
 	}
 	got := e.Error()
 	expect := "invalid address foo, reason: bar"
-	if got != expect {
-		t.Errorf("Expected %v, got %v", expect, got)
-	}
+	require.Equal(t, expect, got)
 	require.ErrorIs(t, ErrInvalidAddr{"foo", "bar"}, e)
 }
 
@@ -26,9 +24,7 @@ func Test_ErrInvalidNumberOfArgs(t *testing.T) {
 	}
 	got := e.Error()
 	expect := "invalid number of arguments; expected 2; got: 1"
-	if got != expect {
-		t.Errorf("Expected %v, got %v", expect, got)
-	}
+	require.Equal(t, expect, got)
 	require.ErrorIs(t, ErrInvalidNumberOfArgs{1, 2}, e)
 }
 
@@ -38,9 +34,7 @@ func Test_ErrInvalidArgument(t *testing.T) {
 	}
 	got := e.Error()
 	expect := "invalid argument: foo"
-	if got != expect {
-		t.Errorf("Expected %v, got %v", expect, got)
-	}
+	require.Equal(t, expect, got)
 	require.ErrorIs(t, ErrInvalidArgument{"foo"}, e)
 }
 
@@ -50,9 +44,7 @@ func Test_ErrInvalidMethod(t *testing.T) {
 	}
 	got := e.Error()
 	expect := "invalid method: foo"
-	if got != expect {
-		t.Errorf("Expected %v, got %v", expect, got)
-	}
+	require.Equal(t, expect, got)
 	require.ErrorIs(t, ErrInvalidMethod{"foo"}, e)
 }
 
@@ -65,9 +57,7 @@ func Test_ErrInvalidCoin(t *testing.T) {
 	}
 	got := e.Error()
 	expect := "invalid coin: denom: foo, is negative: true, is nil: false, is empty: false"
-	if got != expect {
-		t.Errorf("Expected %v, got %v", expect, got)
-	}
+	require.Equal(t, expect, got)
 	require.ErrorIs(t, ErrInvalidCoin{"foo", true, false, false}, e)
 }
 
@@ -77,9 +67,7 @@ func Test_ErrInvalidAmount(t *testing.T) {
 	}
 	got := e.Error()
 	expect := "invalid token amount: foo"
-	if got != expect {
-		t.Errorf("Expected %v, got %v", expect, got)
-	}
+	require.Equal(t, expect, got)
 	require.ErrorIs(t, ErrInvalidAmount{"foo"}, e)
 }
 
@@ -90,9 +78,7 @@ func Test_ErrUnexpected(t *testing.T) {
 	}
 	got := e.Error()
 	expect := "unexpected error in foo: bar"
-	if got != expect {
-		t.Errorf("Expected %v, got %v", expect, got)
-	}
+	require.Equal(t, expect, got)
 	require.ErrorIs(t, ErrUnexpected{"foo", "bar"}, e)
 }
 
@@ -103,9 +89,7 @@ func Test_ErrInsufficientBalance(t *testing.T) {
 	}
 	got := e.Error()
 	expect := "insufficient balance: requested foo, current bar"
-	if got != expect {
-		t.Errorf("Expected %v, got %v", expect, got)
-	}
+	require.Equal(t, expect, got)
 	require.ErrorIs(t, ErrInsufficientBalance{"foo", "bar"}, e)
 }
 
@@ -116,8 +100,16 @@ func Test_ErrInvalidToken(t *testing.T) {
 	}
 	got := e.Error()
 	expect := "invalid token foo: bar"
-	if got != expect {
-		t.Errorf("Expected %v, got %v", expect, got)
-	}
+	require.Equal(t, expect, got)
 	require.ErrorIs(t, ErrInvalidToken{"foo", "bar"}, e)
+}
+
+func Test_ErrWriteMethod(t *testing.T) {
+	e := ErrWriteMethod{
+		Method: "foo",
+	}
+	got := e.Error()
+	expect := "method not allowed in read-only mode: foo"
+	require.Equal(t, expect, got)
+	require.ErrorIs(t, ErrWriteMethod{"foo"}, e)
 }